+++ src/main/java/duke/DeadlineTest.java
@@ -1,4 +1,5 @@
 package duke;
+import duke.task.Deadline;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
