+++ src/main/java/duke/TaskList.java
@@ -1,5 +1,7 @@
 package duke;
 
+import duke.task.Task;
+
 import java.util.ArrayList;
 
 /** A class which encapsulates the list of
