+++ src/main/java/duke/command/FindCommand.java
@@ -1,4 +1,8 @@
-package duke;
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
 
 public class FindCommand extends Command {
